restore former module structure.
