+++ src/cheese.js
@@ -1,0 +1,8 @@
+'use strict';
+const cheeseName = require('cheese-name');
+
+const randomCheese = async (bot, message) => {
+  await bot.reply(message, cheeseName());
+};
+
+module.exports = randomCheese;
